+++ wbooks/src/app/index.tsx
@@ -1,18 +1,13 @@
 import React from 'react';
-import { FlatList, SafeAreaView, StatusBar, ListRenderItem } from 'react-native';
-import BookCell from '@components/BookCell';
-import { Book } from '@interfaces/book';
-import { BOOKS_MOCK } from '@constants/mockBooks';
+import { StatusBar } from 'react-native';
+
+import BookList from './screens/BookList';
 
 const App = () => {
-  const keyExtractor = ({ id }: Book) => String(id);
-  const renderBook: ListRenderItem<Book> = ({ item }) => <BookCell {...item} />;
   return (
     <>
       <StatusBar />
-      <SafeAreaView>
-        <FlatList data={BOOKS_MOCK} renderItem={renderBook} keyExtractor={keyExtractor} />
-      </SafeAreaView>
+      <BookList />
     </>
   );
 };
